checkpoint verification for all tests containing direct checkpoint query
